separate page for resume, coverletter Resumes and Coverletters tabs next to Vacancies Some dummy content on each page Separate widget page for each one
